# src/main/resources/openapi/secretaria-api-v1.yaml
openapi: 3.0.3

info:
  title: SPB Secretaría API
  description: |
    API REST para la gestión académica de la secretaría escolar.
    Versión inicial (v1): centralización de información académica por período.
  version: 1.0.0
  contact:
    name: MaestriaSoft
    email: soporte@maestriasoft.com

servers:
  - url: http://localhost:8080/api/v1
    description: Servidor de desarrollo local

tags:
  - name: informacion-academica
    description: Operaciones sobre la fuente central de información académica por período

paths:

  /periodos/{codigoPeriodo}/informacion-academica:

    post:
      tags:
        - informacion-academica
      operationId: cargarInformacionAcademica
      summary: Cargar información académica para un período
      description: |
        Carga un archivo CSV como fuente central de información académica para el período
        indicado. Si ya existe una fuente central para ese período, es reemplazada
        completamente por la nueva versión. La operación es atómica: si la validación
        o el almacenamiento fallan, la fuente anterior permanece intacta.

        **Formato del archivo CSV:**
        - Codificación: UTF-8
        - Delimitador: punto y coma (`;`)
        - Primera fila: encabezados obligatorios
        - Encabezados: `codigo_estudiante;apellido_paterno;apellido_materno;nombres;calificacion_final;condicion`
        - `calificacion_final`: número decimal entre 0.0 y 20.0
        - `condicion`: `PROMOVIDO`, `REPROBADO` o `ABANDERADO`
      parameters:
        - $ref: '#/components/parameters/CodigoPeriodo'
        - $ref: '#/components/parameters/UsuarioResponsable'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - archivo
              properties:
                archivo:
                  type: string
                  format: binary
                  description: Archivo CSV con los datos académicos del período
      responses:
        '201':
          description: Información académica cargada y registrada como fuente central exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CargaResponse'
        '400':
          description: Archivo inválido — formato incorrecto, datos incompletos o reglas de validación no cumplidas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                archivoVacio:
                  summary: Archivo vacío
                  value:
                    codigo: ARCHIVO_VACIO
                    mensaje: El archivo no puede estar vacío
                encabezadosInvalidos:
                  summary: Encabezados incorrectos
                  value:
                    codigo: ENCABEZADOS_INVALIDOS
                    mensaje: "Los encabezados del archivo no coinciden con el formato esperado. Esperado: codigo_estudiante;apellido_paterno;apellido_materno;nombres;calificacion_final;condicion"
                calificacionInvalida:
                  summary: Calificación fuera de rango
                  value:
                    codigo: CALIFICACION_INVALIDA
                    mensaje: "Fila 5: calificacion_final '25.5' no es válida. Debe ser un número entre 0.0 y 20.0"
                headerAusente:
                  summary: Header X-Usuario-Responsable ausente
                  value:
                    codigo: HEADER_REQUERIDO
                    mensaje: "El header 'X-Usuario-Responsable' es obligatorio"
        '404':
          description: Período académico no encontrado o no habilitado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                periodoNoEncontrado:
                  summary: Período no existe
                  value:
                    codigo: PERIODO_NO_ENCONTRADO
                    mensaje: "El período '2025-III' no existe en el sistema"
                periodoNOHabilitado:
                  summary: Período cerrado
                  value:
                    codigo: PERIODO_NO_HABILITADO
                    mensaje: "El período '2024-I' está cerrado y no acepta cargas de información"
        '413':
          description: Archivo demasiado grande (supera el límite de 10 MB)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      tags:
        - informacion-academica
      operationId: consultarInformacionAcademica
      summary: Consultar la fuente central de información de un período
      description: |
        Devuelve el estado y los metadatos de la fuente central de información académica
        para el período indicado. Si no existe información cargada para el período,
        devuelve estado SIN_INFORMACION con un mensaje orientativo.
      parameters:
        - $ref: '#/components/parameters/CodigoPeriodo'
      responses:
        '200':
          description: Estado de la fuente central recuperado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FuenteCentralResponse'
              examples:
                conInformacion:
                  summary: Período con información cargada
                  value:
                    codigoPeriodo: "2025-II"
                    estado: "DISPONIBLE"
                    nombreArchivo: "datos_academicos_2025II.csv"
                    tamanioBytes: 45312
                    fechaCarga: "2026-01-15T10:30:00"
                    usuarioResponsable: "secretaria01"
                    mensaje: "Disponible para revisión"
                sinInformacion:
                  summary: Período sin información
                  value:
                    codigoPeriodo: "2025-I"
                    estado: "SIN_INFORMACION"
                    nombreArchivo: null
                    tamanioBytes: null
                    fechaCarga: null
                    usuarioResponsable: null
                    mensaje: "No existe fuente central para este período. Realice una carga de información."
        '404':
          description: Período académico no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:

  parameters:
    CodigoPeriodo:
      name: codigoPeriodo
      in: path
      required: true
      description: Código único del período académico (ej. `2025-II`)
      schema:
        type: string
        maxLength: 10
        example: "2025-II"

    UsuarioResponsable:
      name: X-Usuario-Responsable
      in: header
      required: true
      description: Identificador del usuario que realiza la operación de carga (login de la secretaria)
      schema:
        type: string
        maxLength: 100
        example: "secretaria01"

  schemas:

    CargaResponse:
      type: object
      required:
        - codigoPeriodo
        - nombreArchivo
        - tamanioBytes
        - fechaCarga
        - usuarioResponsable
        - mensaje
      properties:
        codigoPeriodo:
          type: string
          description: Código del período para el que se realizó la carga
          example: "2025-II"
        nombreArchivo:
          type: string
          description: Nombre original del archivo cargado
          example: "datos_academicos_2025II.csv"
        tamanioBytes:
          type: integer
          format: int64
          description: Tamaño del archivo en bytes
          example: 45312
        fechaCarga:
          type: string
          format: date-time
          description: Fecha y hora en que se registró la carga
          example: "2026-01-15T10:30:00"
        usuarioResponsable:
          type: string
          description: Usuario que realizó la carga
          example: "secretaria01"
        mensaje:
          type: string
          description: Mensaje de confirmación
          example: "Información académica del período 2025-II registrada exitosamente como fuente central."

    FuenteCentralResponse:
      type: object
      required:
        - codigoPeriodo
        - estado
        - mensaje
      properties:
        codigoPeriodo:
          type: string
          description: Código del período consultado
          example: "2025-II"
        estado:
          type: string
          enum:
            - DISPONIBLE
            - SIN_INFORMACION
          description: Estado de la fuente central para el período
        nombreArchivo:
          type: string
          nullable: true
          description: Nombre del archivo cargado (null si no hay información)
          example: "datos_academicos_2025II.csv"
        tamanioBytes:
          type: integer
          format: int64
          nullable: true
          description: Tamaño del archivo en bytes (null si no hay información)
          example: 45312
        fechaCarga:
          type: string
          format: date-time
          nullable: true
          description: Fecha y hora de la última carga (null si no hay información)
          example: "2026-01-15T10:30:00"
        usuarioResponsable:
          type: string
          nullable: true
          description: Usuario que realizó la última carga (null si no hay información)
          example: "secretaria01"
        mensaje:
          type: string
          description: Mensaje descriptivo del estado para la secretaria
          example: "Disponible para revisión"

    ErrorResponse:
      type: object
      required:
        - codigo
        - mensaje
      properties:
        codigo:
          type: string
          description: Código interno del error
          example: "ARCHIVO_VACIO"
        mensaje:
          type: string
          description: Descripción del error orientada al usuario
          example: "El archivo no puede estar vacío"
        detalle:
          type: string
          nullable: true
          description: Información técnica adicional (opcional)